+++ data/embeds.js
@@ -1,0 +1,10 @@
+module.exports = {
+    default: {
+        description: ``,
+        color: 7121879,
+    },
+    error: { 
+        description: ``, 
+        color: 16329259 
+    }
+}
