+++ data/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    emotes: require('./emotes'),
+    embeds: require('./embeds'),
+}
