+++ eventhandlers/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    init: (ctx) => {
+        require('./bot').init(ctx)
+        require('./mongoose').init(ctx)
+    }
+}
