+++ eventhandlers/mongoose.js
@@ -1,0 +1,5 @@
+const init = (ctx) => {
+    ctx.mcn.on('error', err => {
+        console.log(err)
+    })
+}
