+++ modules/index.js
@@ -1,3 +1,4 @@
 module.exports = {
-    repost: require('./repost')
+    repost: require('./repost'),
+    admin: require('./admin'),
 }
